pyproject.toml
[tool.poetry]
name = "lms-drf"
version = "0.1.0"
description = ""
authors = ["Dimanskr <dimanskr2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.49"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
